@@ -11,44 +11,44 @@
     "postinstall": "ng build --aot --prod"
   },
   "engines": {
-    "node": "8.11.1",
-    "npm": "5.6.0"
+    "node": "12.4.0",
+    "npm": "6.9.0"
   },
   "private": true,
   "dependencies": {
     "@angular-mdc/web": "^1.0.3",
-    "@angular/animations": "^7.2.1",
+    "@angular/animations": "^8.0.0",
     "@angular/cdk": "^7.2.1",
-    "@angular/cli": "~7.2.2",
-    "@angular/common": "^7.2.1",
-    "@angular/compiler": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.1",
-    "@angular/core": "^7.2.1",
-    "@angular/forms": "^7.2.1",
+    "@angular/cli": "~8.0.2",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/forms": "^8.0.0",
     "@angular/http": "^7.2.1",
     "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^7.2.1",
-    "@angular/platform-browser-dynamic": "^7.2.1",
-    "@angular/router": "^7.2.1",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "aws-sdk": "^2.389.0",
     "classlist.js": "^1.1.20150312",
     "core-js": "^2.5.4",
     "cors": "^2.8.5",
     "dotenv": "^6.2.0",
-    "rxjs": "~6.3.3",
-    "rxjs-compat": "^6.3.3",
-    "typescript": "~3.2.4",
+    "node-sass": "^4.12.0",
+    "rxjs": "^6.5.2",
+    "typescript": "~3.4.5",
     "web-animations-js": "^2.3.1",
-    "zone.js": "~0.8.28"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular/language-service": "^7.2.1",
+    "@angular-devkit/build-angular": "^0.800.2",
+    "@angular/language-service": "^8.0.0",
     "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
