@@ -1,11 +1,6 @@
 import { NgModule } from '@angular/core';
 import {MatIconModule} from '@angular/material/icon';
-import {MatSnackBarModule} from '@angular/material/snack-bar';
-import {MatMenuModule} from '@angular/material/menu';
-import {MatDatepickerModule} from '@angular/material/datepicker';
 import {MatNativeDateModule} from '@angular/material/';
-import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
-import {MatTableModule} from '@angular/material/table';
 import {MatPaginatorModule} from '@angular/material/paginator';
 import {MatSliderModule} from '@angular/material/slider';
 import {MatRippleModule} from '@angular/material/core';
@@ -21,6 +16,9 @@ import { MdcCheckboxModule } from '@angular-mdc/web';
 import { MdcButtonModule } from '@angular-mdc/web';
 import { MdcIconModule } from '@angular-mdc/web';
 import { MdcCardModule } from '@angular-mdc/web';
+import { MdcIconButtonModule } from '@angular-mdc/web';
+import { MdcTypographyModule } from '@angular-mdc/web';
+import { MdcElevationModule } from '@angular-mdc/web';
 
 
 
@@ -30,16 +28,13 @@ import { MdcCardModule } from '@angular-mdc/web';
 
 @NgModule({
   imports: [MdcCardModule, MatDividerModule, MdcButtonModule, MatIconModule, MdcIconModule,
-            MatTabsModule, MatSnackBarModule, MatMenuModule,
-            MatDatepickerModule, MatNativeDateModule, MatProgressSpinnerModule,
-            MatTableModule, MatPaginatorModule, MatSliderModule, MatRippleModule,
+            MatTabsModule,  MdcIconButtonModule, MdcTypographyModule, MatNativeDateModule
+            , MatPaginatorModule, MatSliderModule, MatRippleModule, MdcElevationModule,
             BrowserAnimationsModule, MatTooltipModule, MatCardModule, MatBadgeModule, MdcFormFieldModule, MdcCheckboxModule],
 
-  exports: [MdcCardModule, MatDividerModule, MdcButtonModule, MatIconModule,
-             MatTabsModule, MdcIconModule,
-             MatSnackBarModule, MatMenuModule,
-             MatDatepickerModule, MatProgressSpinnerModule, MatTableModule, MatPaginatorModule,
-             MatSliderModule, MatRippleModule, BrowserAnimationsModule,
+  exports: [MdcCardModule, MatDividerModule, MdcButtonModule, MatIconModule, MatNativeDateModule,
+             MatTabsModule, MdcIconModule, MdcIconButtonModule, MdcTypographyModule, MatPaginatorModule,
+             MatSliderModule, MatRippleModule, BrowserAnimationsModule, MdcElevationModule,
              MatTooltipModule, MatCardModule, MatBadgeModule, MdcFormFieldModule, MdcCheckboxModule],
 })
 export class MaterialModule {}
